Change the deploy directory and replace legacy site
Changed the deploy directory from com_jubalrife/dist to ../ so now when the vite build happens, the out directory will be the deployed directory.

Removed the legacy webpage that was living in ../ before this change.

Todo: I will probably swap this to a separate branch deploy at some point in the future